+++ src/Helpers/InternalContainer.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Pouch\Helpers;
+
+interface InternalContainer
+{
+}