+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2
+
+### Changed
+
+- The `get` method is now the main container fetching method instead of resolve.
+- The `resolve` method in now used to go one step further if necessary and call `getContent()` if we're dealing with an internal container.
+
 ## 1.1.1 - 2019-03-11
 
 ## Added